Move all creation logic into ctors
