+++ pkg/secretless/config/v2/config.go
@@ -5,6 +5,7 @@ package v2
 
 import (
 	"github.com/cyberark/secretless-broker/pkg/secretless/config/v1"
+	"gopkg.in/yaml.v2"
 	"sort"
 )
 
@@ -20,13 +21,15 @@ type Service struct {
 	Config      []byte
 }
 
-func NewV1Config(fileContents []byte) (*v1.Config, error) {
-	v2cfg, err := NewConfig(fileContents)
+// NewV1Config is converts the bytes of a v2 YAML file to a v1.Config.  As such,
+// it's the primary public interface of the v2 package.
+func NewV1Config(v2YAML []byte) (*v1.Config, error) {
+	v2cfg, err := NewConfig(v2YAML)
 	if err != nil {
 		return nil, err
 	}
 
-	v1cfg, err := v2cfg.ConvertToV1()
+	v1cfg, err := NewV1ConfigFromV2Config(v2cfg)
 	if err != nil {
 		return nil, err
 	}
@@ -34,22 +37,13 @@ func NewV1Config(fileContents []byte) (*v1.Config, error) {
 	return v1cfg, nil
 }
 
-func NewConfig(fileContents []byte) (*Config, error) {
-	cfgYAML, err := NewConfigYAML(fileContents)
-	if err != nil {
-		return nil, err
-	}
-
-	return cfgYAML.ConvertToConfig()
-}
-
-func (cfg *Config) ConvertToV1() (*v1.Config, error) {
+func NewV1ConfigFromV2Config(v2cfg *Config) (*v1.Config, error) {
 	v1Config := &v1.Config{
 		Listeners: make([]v1.Listener, 0),
 		Handlers:  make([]v1.Handler, 0),
 	}
 
-	for _, svc := range cfg.Services {
+	for _, svc := range v2cfg.Services {
 		v1Svc, err := NewV1Service(*svc)
 		if err != nil {
 			return nil, err
@@ -69,4 +63,52 @@ func (cfg *Config) ConvertToV1() (*v1.Config, error) {
 	})
 
 	return v1Config, nil
+}
+
+func NewConfig(fileContents []byte) (*Config, error) {
+	cfgYAML, err := NewConfigYAML(fileContents)
+	if err != nil {
+		return nil, err
+	}
+
+	return NewConfigFromConfigYAML(cfgYAML)
+}
+
+func NewConfigFromConfigYAML(cfgYAML *configYAML) (*Config, error) {
+	services := make([]*Service, 0)
+	for svcName, svcYAML := range cfgYAML.Services {
+		svc, err := NewService(svcName, svcYAML)
+		if err != nil {
+			return nil, err
+		}
+		services = append(services, svc)
+	}
+
+	return &Config{
+		Services: services,
+	}, nil
+}
+
+func NewService(svcName string, svcYAML *serviceYAML) (*Service, error) {
+
+	credentials, err := NewCredentials(svcYAML.Credentials)
+	if err != nil {
+		return nil, err
+	}
+
+	svc := &Service{
+		Name:        svcName,
+		Credentials: credentials,
+		Protocol:    svcYAML.Protocol,
+		ListenOn:    svcYAML.ListenOn,
+		Config:      nil,
+	}
+
+	configBytes, err := yaml.Marshal(svcYAML.Config)
+	if err != nil {
+		return nil, err
+	}
+	svc.Config = configBytes
+
+	return svc, nil
 }
