+++ pkg/secretless/config/v2/v1_service.go
@@ -1,11 +1,11 @@
 /*
 
 v1Service exists for conceptual clarity.  The concept of a service exists
-implicitly in v1 config.Config, but not in the code.  The combination of a
-Listener and Handler implicitly represents a service in the v1 code. We're
-making it explicit here.
+implicitly in v1.Config, but not in the code.  The combination of a Listener and
+Handler implicitly represents a service in the v1 code. We're making it explicit
+here.
 
-v1Service also hourses protocol specific configuration logic.  In the future,
+v1Service also houses protocol specific configuration logic.  In the future,
 this logic will be the responsibility of individual v2 services.  We're pulling
 them out now into their own functions both to clarify that this is a separate
 step of the configuration process -- one specific to each protocol -- and to
